Fix login rate limit semantics
