+++ src/middlewares/mod.rs
@@ -1,4 +1,5 @@
 pub mod client_allowlist;
+pub mod rate_limit;
 pub mod tracing;
 
 pub use client_allowlist::ClientAllowlistMiddleware;
